Show the Dynamic Island compact trailing value again
The compact trailing chronometer used a custom DiscreteFormatStyle. A
Live Activity view is archived and rendered by the system out of
process, and a custom format style cannot be archived, so the slot came
up empty.

Go back to the system timer text and let it scale down to fit the slot
instead. Dropping the seconds is not an option: the format styles that
hide them either spell the units out ("2 hours, 20 minutes") or are the
custom kind that cannot be archived.

Split HADynamicIslandView.swift one type per file and add snapshot
tests for the compact trailing slot.
